add basic_set
