+++ tests/inference/test_example_openai_query_http_requests.py
@@ -17,29 +17,7 @@
 import subprocess
 import pytest
 import os
-
-
-def start_serve(model_name):
-    print("start_serve")
-    current_path = os.path.dirname(os.path.abspath(__file__))
-
-    config_path = os.path.join(
-        current_path, "../../.github/workflows/config/" + model_name + "-ci.yaml"
-    )
-
-    os.path.join(current_path, "../../inference/serve.py")
-
-    cmd_serve = ["llm_on_ray-serve", "--config_file", config_path]
-
-    result_serve = subprocess.run(cmd_serve, capture_output=True, text=True)
-
-    # Ensure there are no errors in the serve script execution
-    assert result_serve.returncode == 0, print(
-        "\n" + "Serve error stderr message: " + "\n", result_serve.stderr
-    )
-
-    # Print the output of subprocess.run for checking if output is expected
-    print("\n" + "Serve message: " + "\n", result_serve.stdout)
+from basic_set import start_serve
 
 
 def script_with_args(model_name, streaming_response, max_new_tokens, temperature, top_p):
