+++ tests/inference/test_example_query_openai_sdk.py
@@ -17,38 +17,7 @@
 import subprocess
 import pytest
 import os
-
-os.environ["no_proxy"] = "localhost,127.0.0.1"
-os.environ["OPENAI_API_BASE"] = "http://localhost:8000/v1"
-os.environ["OPENAI_API_KEY"] = "YOUR_OPEN_AI_KEY"
-os.environ["OPENAI_BASE_URL"] = "http://localhost:8000/v1"
-
-
-def start_serve(api_base, model_name):
-    # Other OpenAI SDK tests
-    if api_base != "http://localhost:8000/v1":
-        os.environ["OPENAI_API_BASE"] = api_base
-        os.environ["OPENAI_BASE_URL"] = api_base
-
-    current_path = os.path.dirname(os.path.abspath(__file__))
-
-    config_path = os.path.join(
-        current_path, "../../.github/workflows/config/" + model_name + "-ci.yaml"
-    )
-
-    os.path.join(current_path, "../../inference/serve.py")
-
-    cmd_serve = ["llm_on_ray-serve", "--config_file", config_path]
-
-    result_serve = subprocess.run(cmd_serve, capture_output=True, text=True)
-
-    # Ensure there are no errors in the serve script execution
-    assert result_serve.returncode == 0, print(
-        "\n" + "Serve error stderr message: " + "\n", result_serve.stderr
-    )
-
-    # Print the output of subprocess.run for checking if output is expected
-    print("\n" + "Serve message: " + "\n", result_serve.stdout)
+from basic_set import start_serve
 
 
 def script_with_args(api_base, model_name, streaming_response, max_new_tokens, temperature, top_p):
