+++ tests/requirements.txt
@@ -1,4 +1,3 @@
 pytest
 openai>1.0
 pytest-asyncio
-pydantic==2.7.0
